feat: implement User Stories 1 & 2 - Multi-dimensional aggregations and window functions
Complete implementation of User Story 1 (Multi-Dimensional Sales Analysis) and
User Story 2 (Time-Series Trend Analysis) following Benchmark-Driven Development.

## User Story 1: Multi-Dimensional Sales Analysis (T050-T061) ✅

### Benchmarks (T050-T054)
- Revenue by region and year aggregations (p95 <2s SLA)
- Category performance by quarter with filters (p95 <1s SLA)
- Hierarchical drill-down (year → quarter → month)
- Partition pruning validation (80%+ partition skip)
- Deterministic query results verification

### Implementation (T055-T061)
- QueryPatterns class with revenue_by_dimensions() method
- Multi-dimensional aggregation across time, geography, product, customer
- Drill-down through time hierarchy (drill_down_time_hierarchy())
- Partition pruning demonstration (partition_pruning_comparison())
- Integration with generate CLI for partitioned Parquet output
- Profile mode in analyze CLI with execution plan highlighting
- Query catalog documentation with 3 examples and execution plans

## User Story 2: Time-Series Trend Analysis (T062-T070) ✅

### Benchmarks (T062-T065)
- 3-month moving average (p95 <3s SLA)
- Year-over-year growth calculation (p95 <3s SLA)
- Product rankings by category/quarter (p95 <2s SLA)
- Window function correctness validation

### Implementation (T066-T070)
- Moving average with configurable window size (moving_average_revenue())
- Year-over-year growth with LAG window function (yoy_growth())
- Product rankings with ROW_NUMBER/RANK (product_rankings())
- Window function examples in query catalog (3 patterns)
- Window function highlighting in analyze CLI (--profile option)

## Technical Features

### Query Patterns Module (src/query/patterns.py)
- QueryPatterns class for reusable OLAP query patterns
- Multi-dimensional aggregation with flexible dimensions
- Hierarchical drill-down through time hierarchy
- Partition filter comparison (with/without filters)
- Moving average calculations (3, 6, 12 month windows)
- YoY growth with optional dimension partitioning
- Product rankings with configurable partition keys

### Benchmarks (tests/benchmarks/)
- test_aggregations.py: 8 benchmark tests for US1
- test_window_functions.py: 9 benchmark tests for US2
- Integration tests for correctness validation
- All benchmarks passing with test dataset

### Documentation (docs/query-catalog.md)
- 6 query pattern examples with SQL and expected output
- Execution plan analysis and optimization insights
- Performance summary table with SLA targets
- Keywords for analyzing EXPLAIN output

### CLI Enhancements
- analyze CLI: Added --profile option for execution plan analysis
- Window function highlighting (yellow) in execution plans
- Partition filter highlighting (green) in execution plans

## Test Results

All benchmarks passing on test dataset (1,000 transactions):
- Revenue aggregations: ~10ms average execution time
- Moving average: ~5ms average execution time
- Partition pruning: Demonstrates filter effectiveness
- Deterministic results: 100% consistency across runs

## Bug Fixes

- Fixed latitude/longitude type conversion in geography generator
  - Faker returns mixed types; now explicitly cast to float
  - Prevents PyArrow conversion errors in Parquet writes

## Files Changed

### New Files
- src/query/patterns.py (453 lines)
- tests/benchmarks/test_aggregations.py (206 lines)
- tests/benchmarks/test_window_functions.py (246 lines)
- docs/query-catalog.md (459 lines)

### Modified Files
- src/query/__init__.py (export QueryPatterns)
- src/cli/analyze.py (added --profile option with highlighting)
- src/datagen/generator.py (fixed lat/long type conversion)
- specs/001-olap-core-demo/tasks.md (marked T050-T070 complete)

Total: 1,364 lines of new code + documentation

## Benchmark-Driven Development Workflow

✅ Phase 1: Write benchmarks (T050-T054, T062-T065)
✅ Phase 2: Verify benchmarks initially fail
✅ Phase 3: Implement functionality (T055-T061, T066-T070)
✅ Phase 4: Verify benchmarks now pass

Both User Stories 1 and 2 are fully functional and independently testable.

+++ src/datagen/generator.py
@@ -165,8 +165,8 @@ def generate_dim_geography(
 
                 # Generate realistic lat/long based on country
                 if country['code'] == 'US':
-                    latitude = generator.faker.latitude()
-                    longitude = generator.faker.longitude()
+                    latitude = float(generator.faker.latitude())
+                    longitude = float(generator.faker.longitude())
                 elif country['code'] == 'GB':
                     latitude = random.uniform(50.0, 58.0)
                     longitude = random.uniform(-5.0, 2.0)
@@ -180,8 +180,8 @@ def generate_dim_geography(
                     'region': region,
                     'country': country['name'],
                     'country_code': country['code'],
-                    'latitude': round(latitude, 6),
-                    'longitude': round(longitude, 6),
+                    'latitude': float(round(latitude, 6)),
+                    'longitude': float(round(longitude, 6)),
                     'population_segment': random.choice(population_segments),
                     'timezone': country['timezone'],
                 }
